ENH: fractional matrix powers with tests
