+++ scipy/linalg/matfuncs.py
@@ -6,7 +6,7 @@ from __future__ import division, print_function, absolute_import
 
 __all__ = ['expm','expm2','expm3','cosm','sinm','tanm','coshm','sinhm',
            'tanhm','logm','funm','signm','sqrtm',
-           'expm_frechet']
+           'expm_frechet', 'logm_new', 'fractional_matrix_power']
 
 from numpy import asarray, Inf, dot, eye, diag, exp, \
      product, logical_not, ravel, transpose, conjugate, \
@@ -25,11 +25,24 @@ from .decomp import eig
 from .decomp_svd import orth, svd
 from .decomp_schur import schur, rsf2csf
 from ._expm_frechet import expm_frechet
+from ._matfuncs_sqrtm import sqrtm
+#from ._matfuncs_inv_ssq import logm_new, fractional_matrix_power
 import warnings
 
 eps = np.finfo(float).eps
 feps = np.finfo(single).eps
 
+def logm_new(A):
+    # This fixes some issue with imports;
+    # this function calls onenormest which is in scipy.sparse.
+    import scipy.linalg._matfuncs_inv_ssq
+    return scipy.linalg._matfuncs_inv_ssq.logm_new(A)
+
+def fractional_matrix_power(A, t):
+    # This fixes some issue with imports;
+    # this function calls onenormest which is in scipy.sparse.
+    import scipy.linalg._matfuncs_inv_ssq
+    return scipy.linalg._matfuncs_inv_ssq.fractional_matrix_power(A, t)
 
 def expm(A, q=None):
     """
@@ -364,179 +377,6 @@ def funm(A, func, disp=True):
         return F, err
 
 
-def _almost_diagonal(M, abstol=1e-13):
-    M_abs_off_diagonal = np.absolute(M - np.diag(np.diag(M)))
-    return np.max(M_abs_off_diagonal) < abstol
-
-
-def _inverse_squaring_helper(T0, theta):
-    """
-    A helper function for inverse scaling and squaring for Pade approximation.
-
-    Parameters
-    ----------
-    T0 : (N, N) array_like upper triangular
-        Matrix involved in inverse scaling and squaring.
-    theta : indexable
-        The values theta[1] .. theta[7] must be available.
-        They represent bounds related to Pade approximation, and they depend
-        on the matrix function which is being computed.
-        For example, different values of theta are required for
-        matrix logarithm than for matrix fractional power.
-
-    Returns
-    -------
-    T : (N, N) array_like upper triangular
-        Composition of zero or more matrix square roots of T0.
-    s : non-negative integer
-        Number of square roots taken.
-    m : positive integer
-        The degree of the Pade approximation.
-
-    Notes
-    -----
-    This subroutine appears as a chunk of lines within
-    a couple of published algorithms; for example it appears
-    as lines 4--35 in algorithm (3.1) of [1], and
-    as lines 3--34 in algorithm (4.1) of [2].
-
-    References
-    ----------
-    .. [1] Nicholas J. Higham and Lijing Lin (2013)
-           "An Improved Schur-Pade Algorithm for Fractional Powers
-           of a Matrix and their Frechet Derivatives."
-
-    .. [2] Awad H. Al-Mohy and Nicholas J. Higham (2012)
-           "Improved Inverse Scaling and Squaring Algorithms
-           for the Matrix Logarithm."
-           SIAM Journal on Scientific Computing, 34 (4). C152-C169.
-           ISSN 1095-7197
-
-    """
-    T = T0
-
-    # Find s0, the smallest s such that the spectral radius
-    # of a certain diagonal matrix is at most theta[7].
-    s0 = 0
-    tmp_diag = np.diag(T)
-    while np.max(np.absolute(tmp_diag - 1)) > theta[7]:
-        tmp_diag = np.sqrt(tmp_diag)
-        s0 += 1
-
-    # Take matrix square roots of T.
-    for i in range(s0):
-        T = _sqrtm_triu(T)
-
-    # Flow control in this section is a little odd.
-    # This is because I am translating algorithm descriptions
-    # which have GOTOs in the publication.
-    s = s0
-    k = 0
-    d2 = onenormest_m1_power(T, 2) ** (1/2)
-    d3 = onenormest_m1_power(T, 3) ** (1/3)
-    a2 = max(d2, d3)
-    m = None
-    for i in (1, 2):
-        if a2 <= theta[i]:
-            m = i
-            break
-    while m is None:
-        if s > s0:
-            d3 = onenormest_m1_power(T, 3) ** (1/3)
-        d4 = onenormest_m1_power(T, 4) ** (1/4)
-        a3 = max(d3, d4)
-        if a3 <= theta[7]:
-            j1 = min(i for i in (3, 4, 5, 6, 7) if a3 <= theta[i])
-            if j1 <= 6:
-                m = j1
-                break
-            elif a3 / 2 <= theta[5] and k < 2:
-                k += 1
-                T = _sqrtm_triu(T)
-                s += 1
-                continue
-        d5 = onenormest_m1_power(T, 5) ** (1/5)
-        a4 = max(d4, d5)
-        eta = min(a3, a4)
-        for i in (6, 7):
-            if eta <= theta[i]:
-                m = i
-                break
-        if m is not None:
-            break
-        T = _sqrtm_triu(T)
-        s += 1
-
-    # Return the root matrix, the number of square roots, and the Pade degree.
-    return T, s, m
-
-
-def fractional_power(A, t):
-    """
-    Compute a fractional power of a matrix.
-
-    This uses algorithm (3.1) of [1].
-
-    Parameters
-    ----------
-    A : (N, N) array_like
-        Matrix whose fractional power to evaluate.
-    t : float
-        Fractional power between -1 and 1 exclusive.
-
-    Returns
-    -------
-    fractional_power : (N, N) array_like
-        The fractional power of the matrix.
-
-    References
-    ----------
-    .. [1] Nicholas J. Higham and Lijing Lin (2013)
-           "An Improved Schur-Pade Algorithm for Fractional Powers
-           of a Matrix and their Frechet Derivatives."
-
-    """
-    A = asarray(A)
-    if len(A.shape) != 2:
-        raise ValueError("Non-matrix input to matrix function.")
-    T, Z = schur(A)
-    T, Z = rsf2csf(T,Z)
-    m_to_theta = {
-            1 : 1.51e-5,
-            2 : 2.24e-3,
-            3 : 1.88e-2,
-            4 : 6.04e-2,
-            5 : 1.24e-1,
-            6 : 2.00e-1,
-            7 : 2.79e-1,
-            #8 : 3.55e-1,
-            #9 : 4.25e-1,
-            #10 : 4.87e-1,
-            #11 : 5.42e-1,
-            #12 : 5.90e-1,
-            #13 : 6.32e-1,
-            #14 : 6.69e-1,
-            #15 : 7.00e-1,
-            #16 : 7.28e-1,
-            #32 : 9.15e-1,
-            #64 : 9.76e-1,
-            }
-    if _almost_diagonal(T):
-        U = np.diag(np.diag(T) ** t)
-        U, Z = all_mat(U, Z)
-        X = (Z * U * Z.H)
-        return X.A
-    else:
-        T0 = T
-        T, s, m = _inverse_squaring_helper(T, theta)
-
-
-    U, Z = all_mat(U, Z)
-    X = (Z * U * Z.H)
-    return X.A
-
-
-
 def logm(A, disp=True):
     """
     Compute matrix logarithm.
@@ -589,207 +429,6 @@ def logm(A, disp=True):
         return F
     else:
         return F, errest
-
-
-def logm_new(A):
-    """
-    Compute matrix logarithm.
-
-    The matrix logarithm is the inverse of
-    expm: expm(logm(`A`)) == `A`
-
-    Parameters
-    ----------
-    A : (N, N) array_like
-        Matrix whose logarithm to evaluate
-
-    Returns
-    -------
-    logm : (N, N) ndarray
-        Matrix logarithm of `A`
-
-    References
-    ----------
-    .. [1] Awad H. Al-Mohy and Nicholas J. Higham (2012)
-           "Improved Inverse Scaling and Squaring Algorithms
-           for the Matrix Logarithm."
-           SIAM Journal on Scientific Computing, 34 (4). C152-C169.
-           ISSN 1095-7197
-
-    .. [2] Nicholas J. Higham (2008)
-           "Functions of Matrices: Theory and Computation"
-           ISBN 978-0-898716-46-7
-
-    .. [3] Nicholas J. Higham and Lijing lin (2011)
-           "A Schur-Pade Algorithm for Fractional Powers of a Matrix."
-           SIAM Journal on Matrix Analysis and Applications,
-           32 (3). pp. 1056-1078. ISSN 0895-4798
-
-    """
-    A = asarray(A)
-    if len(A.shape) != 2:
-        raise ValueError("Non-matrix input to matrix function.")
-    T, Z = schur(A)
-    T, Z = rsf2csf(T,Z)
-    T0 = T
-
-    # Define bounds given in Table (2.1).
-    theta = (None,
-            1.59e-5, 2.31e-3, 1.94e-2, 6.21e-2,
-            1.28e-1, 2.06e-1, 2.88e-1, 3.67e-1,
-            4.39e-1, 5.03e-1, 5.60e-1, 6.09e-1,
-            6.52e-1, 6.89e-1, 7.21e-1, 7.49e-1)
-
-    T, s, m = _inverse_squaring_helper(T, theta)
-
-    #TODO line 35
-    #NOTE I am starting to think that fractional matrix powers 
-    #NOTE should be implemented before implementing the matrix logarithm.
-    # Replace diag(T - I) by diag(T0)^(1/(2^s)) - 1
-    # using [1, Alg. 2] and recompute the first superdiagonal of T
-    # using [18 (my [3]), eq. (5, 6)] applied to T0.
-
-    #TODO line 36
-    # Evaluate U = 2**s r_m(T - I) using the partial fraction expansion (1.1).
-
-    # Recompute diagonal entries of U.
-    U[np.diag_indices(n)] = np.log(np.diag(T0))
-
-    # Recompute superdiagonal entries of U.
-    # This indexing of this code should be renovated
-    # when newer np.diagonal() becomes available.
-    for i in range(n-1):
-        l1 = T0[i, i]
-        l2 = T0[i+1, i+1]
-        t12 = T0[i, i+1]
-        U[i, i+1] = _logm_superdiag_entry(l1, l2, t12)
-
-    U, Z = all_mat(U, Z)
-    X = (Z * U * Z.H)
-    return X.A
-
-
-def _fractional_power_superdiag_entry(l1, l2, t12, p):
-    """
-    Compute a superdiagonal entry of a matrix fractional power.
-
-    This is Eq. (5.6) in [1].
-
-    Parameters
-    ----------
-    l1 : complex
-        A diagonal entry of the matrix.
-    l2 : complex
-        A diagonal entry of the matrix.
-    t12 : complex
-        A superdiagonal entry of the matrix.
-    p : float
-        A fractional power.
-
-    Returns
-    -------
-    f12 : complex
-        A superdiagonal entry of the matrix fractional power.
-
-    References
-    ----------
-    .. [1] Nicholas J. Higham and Lijing lin (2011)
-           "A Schur-Pade Algorithm for Fractional Powers of a Matrix."
-           SIAM Journal on Matrix Analysis and Applications,
-           32 (3). pp. 1056-1078. ISSN 0895-4798
-
-    """
-    if l1 == l2:
-        return t12 * p * l1**(p-1)
-    elif abs(l1) < abs(l2) / 2 or abs(l2) < abs(l1) / 2:
-        return t12 * ((l2**p) - (l1**p)) / (l2 - l1)
-    else:
-        # This is Eq. (5.5) in [1].
-        z = (l2 - l1) / (l2 + l1)
-        log_l1 = np.log(l1)
-        log_l2 = np.log(l2)
-        atanh_z = np.atanh(z)
-        tmp_a = t12 * np.exp((p/2)*(log_l2 + log_l1))
-        tmp_b = p * (atanh_z + np.pi * 1j * _unwindk(log_l2 - log_l1))
-        tmp_c = 2 * np.sinh(tmp_b) / (l2 - l1)
-        return tmp_a * tmp_c
-
-
-def _logm_superdiag_entry(l1, l2, t12):
-    """
-    Compute a superdiagonal entry of a matrix logarithm.
-
-    This is Eq. (11.28) in [1].
-
-    Parameters
-    ----------
-    l1 : complex
-        A diagonal entry of the matrix.
-    l2 : complex
-        A diagonal entry of the matrix.
-    t12 : complex
-        A superdiagonal entry of the matrix.
-
-    Returns
-    -------
-    f12 : complex
-        A superdiagonal entry of the matrix logarithm.
-
-    References
-    ----------
-    .. [1] Nicholas J. Higham (2008)
-           "Functions of Matrices: Theory and Computation"
-           ISBN 978-0-898716-46-7
-
-    """
-    if l1 == l2:
-        return t12 / l1
-    elif abs(l1) < abs(l2) / 2 or abs(l2) < abs(l1) / 2:
-        return t12 * (np.log(l2) - np.log(l1)) / (l2 - l1)
-    else:
-        z = (l2 - l1) / (l2 + l1)
-        ua = _unwindk(np.log(l2) - np.log(l1))
-        ub = _unwindk(np.log(1+z) - np.log(1-z))
-        return t12 * (2*np.atanh(z) + 2*np.pi*1j*(ua + ub)) / (l2 - l1)
-
-
-def _unwindk(z):
-    """
-    Compute the scalar unwinding number.
-
-    Uses Eq. (5.3) in [1], and should be equal to (z - log(exp(z)) / (2 pi i).
-    Note that this definition differs in sign from the original definition
-    in equations (5, 6) in [2].  The sign convention is justified in [3].
-
-    Parameters
-    ----------
-    z : complex
-        A complex number.
-
-    Returns
-    -------
-    unwinding_number : integer
-        The scalar unwinding number of z.
-
-    References
-    ----------
-    .. [1] Nicholas J. Higham and Lijing lin (2011)
-           "A Schur-Pade Algorithm for Fractional Powers of a Matrix."
-           SIAM Journal on Matrix Analysis and Applications,
-           32 (3). pp. 1056-1078. ISSN 0895-4798
-
-    .. [2] Robert M. Corless and David J. Jeffrey,
-           "The unwinding number." Newsletter ACM SIGSAM Bulletin
-           Volume 30, Issue 2, June 1996, Pages 28-35.
-
-    .. [3] Russell Bradford and Robert M. Corless and James H. Davenport and
-           David J. Jeffrey and Stephen M. Watt,
-           "Reasoning about the elementary functions of complex analysis"
-           Annals of Mathematics and Artificial Intelligence,
-           36: 303-318, 2002.
-
-    """
-    return int(np.ceil((z.imag - np.pi) / (2*np.pi)))
 
 
 def signm(a, disp=True):
@@ -870,137 +509,3 @@ def signm(a, disp=True):
     else:
         return S0, errest
 
-
-def _sqrtm_triu(T, blocksize=64):
-    """
-    Matrix square root of an upper triangular matrix.
-
-    This is a helper function for `sqrtm` and `logm`.
-
-    Parameters
-    ----------
-    T : (N, N) array_like upper triangular
-        Matrix whose square root to evaluate
-    blocksize : integer, optional
-        If the blocksize is not degenerate with respect to the
-        size of the input array, then use a blocked algorithm. (Default: 64)
-
-    Returns
-    -------
-    sqrtm : (N, N) ndarray
-        Value of the sqrt function at `T`
-
-    References
-    ----------
-    .. [1] Edvin Deadman, Nicholas J. Higham, Rui Ralha (2013)
-           "Blocked Schur Algorithms for Computing the Matrix Square Root,
-           Lecture Notes in Computer Science, 7782. pp. 171-182.
-
-    """
-    n, n = T.shape
-    R = np.zeros((n,n), T.dtype.char)
-    
-    # Take square roots of the diagonal of the triangular matrix.
-    R[np.diag_indices(n)] = np.sqrt(np.diag(T))
-
-    # Compute the number of blocks to use; use at least one block.
-    nblocks = max(n // blocksize, 1)
-
-    # Compute the smaller of the two sizes of blocks that
-    # we will actually use, and compute the number of large blocks.
-    bsmall, nlarge = divmod(n, nblocks)
-    blarge = bsmall + 1
-    nsmall = nblocks - nlarge
-    if nsmall * bsmall + nlarge * blarge != n:
-        raise Exception('internal inconsistency')
-
-    # Define the index range covered by each block.
-    start_stop_pairs = []
-    start = 0
-    for count, size in ((nsmall, bsmall), (nlarge, blarge)):
-        for i in range(count):
-            start_stop_pairs.append((start, start + size))
-            start += size
-
-    # Within-block interactions.
-    for start, stop in start_stop_pairs:
-        for j in range(start, stop):
-            for i in range(j-1, start-1, -1):
-                s = 0
-                if j - i > 1:
-                    s = R[i, i+1:j].dot(R[i+1:j, j])
-                R[i,j] = (T[i,j] - s)/(R[i,i] + R[j,j])
-
-    # Between-block interactions.
-    for j in range(nblocks):
-        jstart, jstop = start_stop_pairs[j]
-        for i in range(j-1, -1, -1):
-            istart, istop = start_stop_pairs[i]
-            S = T[istart:istop, jstart:jstop]
-            if j - i > 1:
-                S = S - R[istart:istop, istop:jstart].dot(
-                        R[istop:jstart, jstart:jstop])
-
-            # Invoke LAPACK.
-            # For more details, see the solve_sylvester implemention
-            # and the fortran ztrsyl docs.
-            Rii = R[istart:istop, istart:istop]
-            Rjj = R[jstart:jstop, jstart:jstop]
-            x, scale, info = ztrsyl(Rii, Rjj, S)
-            R[istart:istop, jstart:jstop] = x * scale
-    
-    # Return the matrix square root.
-    return R
-
-
-def sqrtm(A, disp=True, blocksize=64):
-    """
-    Matrix square root.
-
-    Parameters
-    ----------
-    A : (N, N) array_like
-        Matrix whose square root to evaluate
-    disp : bool, optional
-        Print warning if error in the result is estimated large
-        instead of returning estimated error. (Default: True)
-    blocksize : integer, optional
-        If the blocksize is not degenerate with respect to the
-        size of the input array, then use a blocked algorithm. (Default: 64)
-
-    Returns
-    -------
-    sqrtm : (N, N) ndarray
-        Value of the sqrt function at `A`
-
-    errest : float
-        (if disp == False)
-
-        Frobenius norm of the estimated error, ||err||_F / ||A||_F
-
-    References
-    ----------
-    .. [1] Edvin Deadman, Nicholas J. Higham, Rui Ralha (2013)
-           "Blocked Schur Algorithms for Computing the Matrix Square Root,
-           Lecture Notes in Computer Science, 7782. pp. 171-182.
-
-    """
-    A = asarray(A)
-    if len(A.shape) != 2:
-        raise ValueError("Non-matrix input to matrix function.")
-    if blocksize < 1:
-        raise ValueError("The blocksize should be at least 1.")
-    T, Z = schur(A)
-    T, Z = rsf2csf(T,Z)
-    R = _sqrtm_triu(T, blocksize=blocksize)
-    R, Z = all_mat(R,Z)
-    X = (Z * R * Z.H)
-
-    if disp:
-        nzeig = np.any(diag(T) == 0)
-        if nzeig:
-            print("Matrix is singular and may not have a square root.")
-        return X.A
-    else:
-        arg2 = norm(X*X - A,'fro')**2 / norm(A,'fro')
-        return X.A, arg2
